# Optimize package organization to reduce antctl binary size