--- a/pkg/agent/apiserver/handlers/agentinfo/handler.go
+++ b/pkg/agent/apiserver/handlers/agentinfo/handler.go
@@ -18,26 +18,12 @@ import (
 	"encoding/json"
 	"net/http"
 
-	corev1 "k8s.io/api/core/v1"
 	"k8s.io/klog/v2"
 
+	"antrea.io/antrea/pkg/agent/apis"
 	"antrea.io/antrea/pkg/agent/querier"
-	"antrea.io/antrea/pkg/antctl/transform/common"
 	"antrea.io/antrea/pkg/apis/crd/v1beta1"
 )
-
-// AntreaAgentInfoResponse is the struct for the response of agentinfo command.
-// It includes all fields except meta info from v1beta1.AntreaAgentInfo struct.
-type AntreaAgentInfoResponse struct {
-	Version                     string                              `json:"version,omitempty"`                     // Antrea binary version
-	PodRef                      corev1.ObjectReference              `json:"podRef,omitempty"`                      // The Pod that Antrea Agent is running in
-	NodeRef                     corev1.ObjectReference              `json:"nodeRef,omitempty"`                     // The Node that Antrea Agent is running in
-	NodeSubnets                 []string                            `json:"nodeSubnets,omitempty"`                 // Node subnets
-	OVSInfo                     v1beta1.OVSInfo                     `json:"ovsInfo,omitempty"`                     // OVS Information
-	NetworkPolicyControllerInfo v1beta1.NetworkPolicyControllerInfo `json:"networkPolicyControllerInfo,omitempty"` // Antrea Agent NetworkPolicy information
-	LocalPodNum                 int32                               `json:"localPodNum,omitempty"`                 // The number of Pods which the agent is in charge of
-	AgentConditions             []v1beta1.AgentCondition            `json:"agentConditions,omitempty"`             // Agent condition contains types like AgentHealthy
-}
 
 // HandleFunc returns the function which can handle queries issued by agentinfo commands.
 // The handler function populates Antrea agent information to the response.
@@ -45,7 +31,7 @@ func HandleFunc(aq querier.AgentQuerier) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		agentInfo := new(v1beta1.AntreaAgentInfo)
 		aq.GetAgentInfo(agentInfo, false)
-		info := &AntreaAgentInfoResponse{
+		info := &apis.AntreaAgentInfoResponse{
 			Version:                     agentInfo.Version,
 			PodRef:                      agentInfo.PodRef,
 			NodeRef:                     agentInfo.NodeRef,
@@ -61,41 +47,4 @@ func HandleFunc(aq querier.AgentQuerier) http.HandlerFunc {
 			klog.Errorf("Error when encoding AntreaAgentInfo to json: %v", err)
 		}
 	}
-}
-
-var _ common.TableOutput = new(AntreaAgentInfoResponse)
-
-func (r AntreaAgentInfoResponse) GetTableHeader() []string {
-	return []string{"POD", "NODE", "STATUS", "NODE-SUBNET", "NETWORK-POLICIES", "ADDRESS-GROUPS", "APPLIED-TO-GROUPS", "LOCAL-PODS"}
-}
-
-func (r AntreaAgentInfoResponse) GetAgentConditionStr() string {
-	if r.AgentConditions == nil {
-		return ""
-	}
-	agentCondition := "Healthy"
-	for _, cond := range r.AgentConditions {
-		if cond.Status == corev1.ConditionUnknown {
-			agentCondition = "Unknown"
-		}
-		if cond.Status == corev1.ConditionFalse {
-			return "Unhealthy"
-		}
-	}
-	return agentCondition
-}
-
-func (r AntreaAgentInfoResponse) GetTableRow(maxColumnLength int) []string {
-	return []string{r.PodRef.Namespace + "/" + r.PodRef.Name,
-		r.NodeRef.Name,
-		r.GetAgentConditionStr(),
-		common.GenerateTableElementWithSummary(r.NodeSubnets, maxColumnLength),
-		common.Int32ToString(r.NetworkPolicyControllerInfo.NetworkPolicyNum),
-		common.Int32ToString(r.NetworkPolicyControllerInfo.AddressGroupNum),
-		common.Int32ToString(r.NetworkPolicyControllerInfo.AppliedToGroupNum),
-		common.Int32ToString(r.LocalPodNum)}
-}
-
-func (r AntreaAgentInfoResponse) SortRows() bool {
-	return true
 }