--- a/pkg/antctl/runtime/runtime.go
+++ b/pkg/antctl/runtime/runtime.go
@@ -22,7 +22,7 @@ import (
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
 
-	agentapiserver "antrea.io/antrea/pkg/agent/apiserver"
+	"antrea.io/antrea/pkg/apis"
 	"antrea.io/antrea/pkg/util/runtime"
 )
 
@@ -59,7 +59,7 @@ func init() {
 		strings.HasPrefix(podName, "flow-aggregator"))
 
 	if runtime.IsWindowsPlatform() && !InPod {
-		if _, err := os.Stat(agentapiserver.TokenPath); err == nil {
+		if _, err := os.Stat(apis.APIServerLoopbackTokenPath); err == nil {
 			InPod = true
 			Mode = ModeAgent
 			return