As a client, antctl needs to know some well-known resource names and the structs of the API responses. However, the code placement makes antctl import a lot of packages which are supposed to be consumed by server side code, e.g.  antrea-controller, antrea-agent.

This patch refactors relevant packages, extracts well-known resources names and API structs to separate common packages, mostly apis packages. It reduces antctl binary size by 10MB, from 55MB to 45MB. The placement also makes it more explicit that these constants and structs are API related and should be changed more carefully.

The major changes are as follows:
1. Move well-known CRD resource names to `types.go` files which declare these CRD structs.
2. Move other well-known resource names to package `pkg/apis`.
3. Move structs of API responses to package `apis` of components which provide the corresponding APIs.

For #5883

-------
Before:
```
-rwxr-xr-x  1 root root  55M Feb 29 09:45 antctl

antrea/antrea-controller-ubuntu           v2.0.0-dev-d28e282d0   a171ba592d24   8 seconds ago    197MB
antrea/antrea-agent-ubuntu                v2.0.0-dev-d28e282d0   7e84ada052f8   15 seconds ago   350MB
```
After:
```
-rwxr-xr-x  1 root root  45M Feb 29 09:49 antctl

antrea/antrea-controller-ubuntu           v2.0.0-dev-c94b2155a   15c6096c8ef2   27 seconds ago   187MB
antrea/antrea-agent-ubuntu                v2.0.0-dev-c94b2155a   648ab7d783ca   38 seconds ago   340MB
```